Refactor and optimize smart contracts in beefy_strategies and oracle modules
- Updated Multicall contract for improved readability and consistency in
function formatting.
- Refactored Path library to enhance code clarity and maintainability.
- Modified SkyToNativeSwap contract to use consistent uint256 type for
amounts.
- Standardized amount variable types across SwapperIfNotZero and
SwapperSwap contracts.
- Enhanced TickMath library with clearer error messages and consistent
formatting.
- Cleaned up UniV3Actions library for better readability and reduced
line length.
- Improved UniV4Swapper contract by simplifying function parameters and
enhancing error handling.
- Updated UniswapV3OracleLibrary for consistent error messages and
improved readability.
- Refactored UniswapV3Utils library for better clarity and consistency
in function definitions.
- Cleaned up console library for improved readability and reduced line
length.
- Optimized scUSDTostkscUSD contract for consistent variable types and
improved readability.
- Removed unnecessary whitespace and comments in BeefyOracleErrors and
a BeefyOracleHelper contracts.
- Enhanced HealthOracle contract for better readability and consistent
formatting.
- Refactored VigilZapRouter contract for improved readability and
consistency in function calls.
- Cleaned up SeedToken contract for consistent formatting.
- Updated HealthOracle and VigilZapRouter test cases for improved
 readability and consistency.

+++ apps/contracts/src/beefy_strategies/Aave/StrategyAaveSupplyOnly.sol
@@ -38,17 +38,17 @@ contract StrategyAaveSupplyOnly is BaseAllToNativeFactoryStrat {
         return "Aave";
     }
 
-    function balanceOfPool() public view override returns (uint) {
+    function balanceOfPool() public view override returns (uint256) {
         return storedBalance;
     }
 
-    function _deposit(uint amount) internal override {
+    function _deposit(uint256 amount) internal override {
         storedBalance += amount;
         IERC20(want).forceApprove(lendingPool, amount);
         ILendingPool(lendingPool).deposit(want, amount, address(this), 0);
     }
 
-    function _withdraw(uint amount) internal override {
+    function _withdraw(uint256 amount) internal override {
         if (amount > 0) {
             storedBalance -= amount;
             ILendingPool(lendingPool).withdraw(want, amount, address(this));
@@ -58,7 +58,7 @@ contract StrategyAaveSupplyOnly is BaseAllToNativeFactoryStrat {
     function _emergencyWithdraw() internal override {
         storedBalance = 0;
         if (IERC20(aToken).balanceOf(address(this)) > 0) {
-            ILendingPool(lendingPool).withdraw(want, type(uint).max, address(this));
+            ILendingPool(lendingPool).withdraw(want, type(uint256).max, address(this));
         }
     }
