+++ apps/contracts/src/beefy_strategies/Balancer/BeefyBalancerStructs.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.0; 
+pragma solidity ^0.8.0;
 
 library BeefyBalancerStructs {
     struct BatchSwapStruct {
@@ -9,13 +9,13 @@ library BeefyBalancerStructs {
     }
 
     struct Reward {
-        mapping(uint => BatchSwapStruct) swapInfo;
+        mapping(uint256 => BatchSwapStruct) swapInfo;
         address[] assets;
         bytes routeToNative; // backup route in case there is no Balancer liquidity for reward
-        uint minAmount; // minimum amount to be swapped to native
+        uint256 minAmount; // minimum amount to be swapped to native
     }
 
-     struct Input {
+    struct Input {
         address input;
         bool isComposable;
         bool isBeets;
