+++ apps/contracts/src/beefy_strategies/Common/BaseAllToNativeStrat.sol
@@ -10,7 +10,7 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
     using SafeERC20 for IERC20;
 
     address[] public rewards;
-    mapping(address => uint) public minAmounts; // tokens minimum amount to be swapped
+    mapping(address => uint256) public minAmounts; // tokens minimum amount to be swapped
 
     address public want;
     address public native;
@@ -25,12 +25,17 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
     event Withdraw(uint256 tvl);
     event ChargedFees(uint256 callFees, uint256 beefyFees, uint256 strategistFees);
 
-    function __BaseStrategy_init(address _want, address _native, address[] calldata _rewards, CommonAddresses calldata _commonAddresses) internal onlyInitializing {
+    function __BaseStrategy_init(
+        address _want,
+        address _native,
+        address[] calldata _rewards,
+        CommonAddresses calldata _commonAddresses
+    ) internal onlyInitializing {
         __StratFeeManager_init(_commonAddresses);
         want = _want;
         native = _native;
 
-        for (uint i; i < _rewards.length; i++) {
+        for (uint256 i; i < _rewards.length; i++) {
             addReward(_rewards[i]);
         }
 
@@ -40,9 +45,9 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
     }
 
     // it calculates how much 'want' the strategy has working in the farm.
-    function balanceOfPool() public view virtual returns (uint);
-    function _deposit(uint amount) internal virtual;
-    function _withdraw(uint amount) internal virtual;
+    function balanceOfPool() public view virtual returns (uint256);
+    function _deposit(uint256 amount) internal virtual;
+    function _withdraw(uint256 amount) internal virtual;
     function _emergencyWithdraw() internal virtual;
     function _claim() internal virtual;
     function _verifyRewardToken(address token) internal view virtual;
@@ -122,7 +127,7 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
     }
 
     function _swapRewardsToNative() internal virtual {
-        for (uint i; i < rewards.length; ++i) {
+        for (uint256 i; i < rewards.length; ++i) {
             address token = rewards[i];
             uint256 amount = IERC20(token).balanceOf(address(this));
             if (amount > minAmounts[token]) {
@@ -160,11 +165,11 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
     }
 
     function _swap(address tokenFrom, address tokenTo) internal {
-        uint bal = IERC20(tokenFrom).balanceOf(address(this));
+        uint256 bal = IERC20(tokenFrom).balanceOf(address(this));
         IBeefySwapper(unirouter).swap(tokenFrom, tokenTo, bal);
     }
 
-    function rewardsLength() external view returns (uint) {
+    function rewardsLength() external view returns (uint256) {
         return rewards.length;
     }
 
@@ -175,41 +180,41 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
 
         rewards.push(_token);
         _approve(_token, unirouter, 0);
-        _approve(_token, unirouter, type(uint).max);
+        _approve(_token, unirouter, type(uint256).max);
     }
 
-    function removeReward(uint i) external onlyManager {
+    function removeReward(uint256 i) external onlyManager {
         rewards[i] = rewards[rewards.length - 1];
         rewards.pop();
     }
 
     function resetRewards() external onlyManager {
-        for (uint i; i < rewards.length; ++i) {
+        for (uint256 i; i < rewards.length; ++i) {
             _approve(rewards[i], unirouter, 0);
         }
         delete rewards;
     }
 
     function updateUnirouter(address _unirouter) external onlyOwner {
-        for (uint i; i < rewards.length; ++i) {
+        for (uint256 i; i < rewards.length; ++i) {
             address token = rewards[i];
             _approve(token, unirouter, 0);
             _approve(token, _unirouter, 0);
-            _approve(token, _unirouter, type(uint).max);
+            _approve(token, _unirouter, type(uint256).max);
         }
         if (depositToken != address(0) && depositToken != native) {
             _approve(depositToken, unirouter, 0);
             _approve(depositToken, _unirouter, 0);
-            _approve(depositToken, _unirouter, type(uint).max);
+            _approve(depositToken, _unirouter, type(uint256).max);
         }
         _approve(native, unirouter, 0);
         _approve(native, _unirouter, 0);
-        _approve(native, _unirouter, type(uint).max);
+        _approve(native, _unirouter, type(uint256).max);
         unirouter = _unirouter;
         emit SetUnirouter(_unirouter);
     }
 
-    function setRewardMinAmount(address token, uint minAmount) external onlyManager {
+    function setRewardMinAmount(address token, uint256 minAmount) external onlyManager {
         minAmounts[token] = minAmount;
     }
 
@@ -223,7 +228,7 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
 
         depositToken = token;
         _approve(token, unirouter, 0);
-        _approve(token, unirouter, type(uint).max);
+        _approve(token, unirouter, type(uint256).max);
     }
 
     function lockedProfit() public view returns (uint256) {
@@ -252,15 +257,15 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
         }
     }
 
-    function setLockDuration(uint _duration) external onlyManager {
+    function setLockDuration(uint256 _duration) external onlyManager {
         lockDuration = _duration;
     }
 
-    function rewardsAvailable() external view virtual returns (uint) {
+    function rewardsAvailable() external view virtual returns (uint256) {
         return 0;
     }
 
-    function callReward() external view virtual returns (uint) {
+    function callReward() external view virtual returns (uint256) {
         return 0;
     }
 
@@ -288,7 +293,7 @@ abstract contract BaseAllToNativeStrat is StratFeeManagerInitializable {
         deposit();
     }
 
-    function _approve(address _token, address _spender, uint amount) internal {
+    function _approve(address _token, address _spender, uint256 amount) internal {
         IERC20(_token).approve(_spender, amount);
     }
 }
